+++ config/dev-local.js
@@ -4,10 +4,13 @@ var devVm  = require('./dev-vm.js')
 
 module.exports = merge(base, devVm, {
 	app: {
+		apolloBatching: false,
 		host: 'localhost',
 		publicPath: '/',
 		photoPath: 'https://www-dev-kiva-org.freetls.fastly.net/img/',
 		graphqlUri: 'https://marketplace-api.dk1.kiva.org/graphql',
+		// Alternate graphql url (marketplace-api.dk1.kiva.org is redirected to this url)
+		// graphqlUri: 'https://gateway.development.kiva.org/graphql',
 		enableAnalytics: false,
 		enableSnowplow: false,
 		snowplowUri: 'events.fivetran.com/snowplow/v5qt54ocr2nm',
@@ -27,6 +30,8 @@ module.exports = merge(base, devVm, {
 	server: {
 		graphqlUri: 'https://marketplace-api.dk1.kiva.org/graphql',
 		sessionUri: 'https://www.dev.kiva.org/start-ui-session',
+		// Alternate session url
+		// sessionUri: 'https://www.development.kiva.org/start-ui-session',
 		memcachedEnabled: false,
 		disableCluster: true,
 	}